test: gitea operator
