+++ package.json
@@ -125,10 +125,6 @@
     "tasks:copy-certs-argo": "node dist/tasks/otomi/copy-certs-argo.js",
     "tasks:drone-dev": "ts-node-dev ./src/tasks/drone/drone.ts",
     "tasks:drone": "node dist/tasks/drone/drone.js",
-    "tasks:gitea-dev": "NODE_TLS_REJECT_UNAUTHORIZED=0 ts-node-dev ./src/tasks/gitea/gitea.ts && ts-node-dev ./src/tasks/gitea/gitea-drone-oauth.ts",
-    "tasks:gitea-drone-auth-dev": "NODE_TLS_REJECT_UNAUTHORIZED=0 ts-node-dev ./src/tasks/gitea/gitea-drone-oauth.ts",
-    "tasks:gitea-drone-auth": "node dist/tasks/gitea/gitea-drone-oauth.js",
-    "tasks:gitea": "node dist/tasks/gitea/gitea.js && node dist/tasks/gitea/gitea-drone-oauth.js",
     "tasks:harbor-dev": "NODE_TLS_REJECT_UNAUTHORIZED=0 ts-node-dev ./src/tasks/harbor/harbor.ts",
     "tasks:harbor": "node dist/tasks/harbor/harbor.js",
     "tasks:keycloak-dev": "NODE_TLS_REJECT_UNAUTHORIZED=0 ts-node-dev ./src/tasks/keycloak/keycloak.ts",
